Final desing ver 1.0
This is the layout i got to came with,
it was a really funny and interesting experience,
the hardest parts where the first sections but
i my improvement over the development, this is
no perfect of, if i get clingy i know there is a
lot of css that can be refactorized and some
html can improve too. But in comparison from my
last website there is SUUREEly huge advances.

Maybe ill be back to this if i feel like to.